spectrum-27: refactor new task to hook
